chore(wallet): centralize descriptor parsing logic in `floresta-watch-only` module
 Improved code organization by consolidating descriptor handling in a single module

- Moved `parse_descriptors`, `parse_xpubs`, and `slip132` to the
`floresta-watch-only` module.
- Centralized descriptor and XPUB parsing logic, previously scattered
across `floresta-node` and `floresta-common`.
- Removed descriptor-related features and the `miniscript` dependency
from `floresta-common`.
