+++ crates/floresta-node/src/lib.rs
@@ -12,7 +12,6 @@ mod error;
 mod florestad;
 #[cfg(feature = "json-rpc")]
 mod json_rpc;
-mod slip132;
 mod wallet_input;
 #[cfg(feature = "zmq-server")]
 mod zmq;
